+++ Example/Pods/Target Support Files/Pods-CZPicker_Tests-CZPicker/Pods-CZPicker_Tests-CZPicker-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_CZPicker_Tests_CZPicker : NSObject
-@end
-@implementation PodsDummy_Pods_CZPicker_Tests_CZPicker
-@end